get sertia catalog

+++ FinalProject/common/src/main/java/org/sertia/contracts/movies/catalog/controller/SertiaMovie.java
@@ -1,9 +1,7 @@
 package org.sertia.contracts.movies.catalog.controller;
 
 public class SertiaMovie extends CinemaScreeningMovie {
-
 	public boolean isStreamable;
-	public int pricePerStream = 0;
+	public double pricePerStream = 0;
 	public boolean isComingSoon;
-
 }
